Refactor code & resolve conflict
