+++ packages/revive/src/executor/EventExecutor.ts
@@ -1,7 +1,7 @@
 import { isEventRecord } from '@dedot/api';
 import { GenericSubstrateApi, IEventRecord, Unsub } from '@dedot/types';
 import { assert, stringPascalCase } from '@dedot/utils';
-import { ContractEvent, ContractEventMeta, GenericContractEvent } from '../types/index.js';
+import { ContractEvent, ContractEventMeta, GenericContractEvent } from '../types.js';
 import { ContractExecutor } from './abstract/index.js';
 
 export class EventExecutor<ChainApi extends GenericSubstrateApi> extends ContractExecutor<ChainApi> {
