+++ packages/revive/src/executor/abstract/ContractExecutor.ts
@@ -2,7 +2,7 @@ import { ISubstrateClient } from '@dedot/api';
 import { AccountId20, AccountId20Like } from '@dedot/codecs';
 import { GenericSubstrateApi } from '@dedot/types';
 import { TypinkRegistry } from '../../TypinkRegistry.js';
-import { ContractCallMessage, ExecutionOptions } from '../../types/index.js';
+import { ContractCallMessage, ExecutionOptions } from '../../types.js';
 import { normalizeLabel } from '../../utils.js';
 import { Executor } from './Executor.js';
