+++ packages/revive/src/index.ts
@@ -1,7 +1,7 @@
 export * from './utils.js';
 export * from './errors.js';
 export * from './executor/index.js';
-export * from './types/index.js';
+export * from './types.js';
 export * from './TypinkRegistry.js';
 export * from './Contract.js';
 export * from './ContractDeployer.js';
